Moved queue command to own class for more functions
